requirements.txt
torch
transformers
datasets
pyyaml
Pillow
tqdm
decord 